@@ -1,22 +1,10 @@
-const baseConfig =
-  process.env.NODE_ENV === "production"
-    ? // when running the Netify CLI or building on Netlify, we want to use
-      {
-        server: "./server.js",
-        serverBuildPath: ".netlify/functions-internal/server.js",
-      }
-    : // otherwise support running remix dev, i.e. no custom server
-      undefined;
+import { config } from "@netlify/remix-adapter";
 
 /** @type {import('@remix-run/dev').AppConfig} */
-module.exports = {
-  ...baseConfig,
-  ignoredRouteFiles: ["**/.*"],
-  // See https://remix.run/docs/en/main/file-conventions/route-files-v2
-  future: {
-    v2_routeConvention: true,
-  }
+export default {
+  ...(process.env.NODE_ENV === "production" ? config : undefined),
+  // This works out of the box with the Netlify adapter, but you can
   // add your own custom config here if you want to.
   //
-  // See https://remix.run/docs/en/v1/file-conventions/remix-config
+  // See https://remix.run/file-conventions/remix-config
 };
