@@ -1,4 +1,5 @@
-import type { MetaFunction } from "@netlify/remix-runtime";
+import { cssBundleHref } from "@remix-run/css-bundle";
+import type { LinksFunction } from "@netlify/remix-runtime";
 import {
   Links,
   LiveReload,
@@ -8,18 +9,16 @@ import {
   ScrollRestoration,
 } from "@remix-run/react";
 
-export const meta: MetaFunction = () => [
-  {
-    charset: "utf-8",
-    title: "New Remix App",
-    viewport: "width=device-width,initial-scale=1",
-  }
+export const links: LinksFunction = () => [
+  ...(cssBundleHref ? [{ rel: "stylesheet", href: cssBundleHref }] : []),
 ];
 
 export default function App() {
   return (
     <html lang="en">
       <head>
+        <meta charSet="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
         <Meta />
         <Links />
       </head>
